refactor: api doc & controller & dto 분리
