+++ src/main/java/in/koreatech/koin/admin/shop/controller/AdminShopApi.java
@@ -14,24 +14,15 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 
-import in.koreatech.koin.admin.shop.dto.AdminCreateMenuCategoryRequest;
-import in.koreatech.koin.admin.shop.dto.AdminCreateMenuRequest;
-import in.koreatech.koin.admin.shop.dto.AdminCreateShopCategoryRequest;
-import in.koreatech.koin.admin.shop.dto.AdminCreateShopRequest;
-import in.koreatech.koin.admin.shop.dto.AdminMenuCategoriesResponse;
-import in.koreatech.koin.admin.shop.dto.AdminMenuDetailResponse;
-import in.koreatech.koin.admin.shop.dto.AdminModifyMenuCategoryRequest;
-import in.koreatech.koin.admin.shop.dto.AdminModifyMenuRequest;
-import in.koreatech.koin.admin.shop.dto.AdminModifyShopCategoriesOrderRequest;
-import in.koreatech.koin.admin.shop.dto.AdminModifyShopCategoryRequest;
-import in.koreatech.koin.admin.shop.dto.AdminModifyShopRequest;
-import in.koreatech.koin.admin.shop.dto.AdminModifyShopReviewReportStatusRequest;
-import in.koreatech.koin.admin.shop.dto.AdminShopCategoryResponse;
-import in.koreatech.koin.admin.shop.dto.AdminShopMenuResponse;
-import in.koreatech.koin.admin.shop.dto.AdminShopParentCategoryResponse;
-import in.koreatech.koin.admin.shop.dto.AdminShopResponse;
-import in.koreatech.koin.admin.shop.dto.AdminShopsResponse;
-import in.koreatech.koin.admin.shop.dto.AdminShopsReviewsResponse;
+import in.koreatech.koin.admin.shop.dto.shop.AdminCreateShopCategoryRequest;
+import in.koreatech.koin.admin.shop.dto.shop.AdminCreateShopRequest;
+import in.koreatech.koin.admin.shop.dto.shop.AdminModifyShopCategoriesOrderRequest;
+import in.koreatech.koin.admin.shop.dto.shop.AdminModifyShopCategoryRequest;
+import in.koreatech.koin.admin.shop.dto.shop.AdminModifyShopRequest;
+import in.koreatech.koin.admin.shop.dto.shop.AdminShopCategoryResponse;
+import in.koreatech.koin.admin.shop.dto.shop.AdminShopParentCategoryResponse;
+import in.koreatech.koin.admin.shop.dto.shop.AdminShopResponse;
+import in.koreatech.koin.admin.shop.dto.shop.AdminShopsResponse;
 import in.koreatech.koin.global.auth.Auth;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -127,88 +118,10 @@ public interface AdminShopApi {
             @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
         }
     )
-    @Operation(summary = "특정 상점의 모든 메뉴 조회")
-    @GetMapping("/admin/shops/{id}/menus")
-    ResponseEntity<AdminShopMenuResponse> getAllMenus(
-        @Parameter(in = PATH) @PathVariable("id") Integer shopId,
-        @Auth(permit = {ADMIN}) Integer adminId
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
-        }
-    )
-    @Operation(summary = "특정 상점의 모든 메뉴 카테고리 조회")
-    @GetMapping("/admin/shops/{id}/menus/categories")
-    ResponseEntity<AdminMenuCategoriesResponse> getAllMenuCategories(
-        @Parameter(in = PATH) @PathVariable("id") Integer shopId,
-        @Auth(permit = {ADMIN}) Integer adminId
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
-        }
-    )
-    @Operation(summary = "특정 상점의 메뉴 조회")
-    @GetMapping("/admin/shops/{shopId}/menus/{menuId}")
-    ResponseEntity<AdminMenuDetailResponse> getMenu(
-        @Parameter(in = PATH) @PathVariable("shopId") Integer shopId,
-        @Parameter(in = PATH) @PathVariable("menuId") Integer menuId,
-        @Auth(permit = {ADMIN}) Integer adminId
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "201"),
-            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
-        }
-    )
-    @Operation(summary = "특정 상점의 메뉴 생성")
-    @PostMapping("/admin/shops/{id}/menus")
-    ResponseEntity<Void> createMenu(
-        @Parameter(in = PATH) @PathVariable("id") Integer shopId,
-        @RequestBody @Valid AdminCreateMenuRequest adminCreateMenuRequest,
-        @Auth(permit = {ADMIN}) Integer adminId
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "201"),
-            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
-        }
-    )
-    @Operation(summary = "특정 상점의 메뉴 카테고리 생성")
-    @PostMapping("/admin/shops/{id}/menus/categories")
-    ResponseEntity<Void> createMenuCategory(
-        @Parameter(in = PATH) @PathVariable("id") Integer shopId,
-        @RequestBody @Valid AdminCreateMenuCategoryRequest adminCreateMenuCategoryRequest,
-        @Auth(permit = {ADMIN}) Integer adminId
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
-        }
-    )
     @Operation(summary = "상점 생성")
     @PostMapping("/admin/shops")
     ResponseEntity<Void> createShop(
-        @RequestBody @Valid AdminCreateShopRequest adminCreateShopRequest,
+        @RequestBody @Valid AdminCreateShopRequest request,
         @Auth(permit = {ADMIN}) Integer adminId
     );
 
@@ -223,7 +136,7 @@ public interface AdminShopApi {
     @Operation(summary = "상점 카테고리 생성")
     @PostMapping("/admin/shops/categories")
     ResponseEntity<Void> createShopCategory(
-        @RequestBody @Valid AdminCreateShopCategoryRequest adminCreateShopCategoryRequest,
+        @RequestBody @Valid AdminCreateShopCategoryRequest request,
         @Auth(permit = {ADMIN}) Integer adminId
     );
 
@@ -254,7 +167,7 @@ public interface AdminShopApi {
     @PutMapping("/admin/shops/{id}")
     ResponseEntity<Void> modifyShop(
         @Parameter(in = PATH) @PathVariable Integer id,
-        @RequestBody @Valid AdminModifyShopRequest adminModifyShopRequest,
+        @RequestBody @Valid AdminModifyShopRequest request,
         @Auth(permit = {ADMIN}) Integer adminId
     );
 
@@ -270,7 +183,7 @@ public interface AdminShopApi {
     @PutMapping("/admin/shops/categories/{id}")
     ResponseEntity<Void> modifyShopCategory(
         @Parameter(in = PATH) @PathVariable Integer id,
-        @RequestBody @Valid AdminModifyShopCategoryRequest adminModifyShopCategoryRequest,
+        @RequestBody @Valid AdminModifyShopCategoryRequest request,
         @Auth(permit = {ADMIN}) Integer adminId
     );
 
@@ -286,40 +199,7 @@ public interface AdminShopApi {
     @Operation(summary = "상점 카테고리 순서 수정")
     @PutMapping("/admin/shops/categories/order")
     ResponseEntity<Void> modifyShopCategoriesOrder(
-        @RequestBody @Valid AdminModifyShopCategoriesOrderRequest adminModifyShopCategoriesOrderRequest,
-        @Auth(permit = {ADMIN}) Integer adminId
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
-        }
-    )
-    @Operation(summary = "특정 상점의 메뉴 카테고리 수정")
-    @PutMapping("/admin/shops/{shopId}/menus/categories")
-    ResponseEntity<Void> modifyMenuCategory(
-        @Parameter(in = PATH) @PathVariable("shopId") Integer shopId,
-        @RequestBody @Valid AdminModifyMenuCategoryRequest adminModifyMenuCategoryRequest,
-        @Auth(permit = {ADMIN}) Integer adminId
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
-        }
-    )
-    @Operation(summary = "특정 상점의 메뉴 수정")
-    @PutMapping("/admin/shops/{shopId}/menus/{menuId}")
-    ResponseEntity<Void> modifyMenu(
-        @Parameter(in = PATH) @PathVariable("shopId") Integer shopId,
-        @Parameter(in = PATH) @PathVariable("menuId") Integer menuId,
-        @RequestBody @Valid AdminModifyMenuRequest adminModifyMenuRequest,
+        @RequestBody @Valid AdminModifyShopCategoriesOrderRequest request,
         @Auth(permit = {ADMIN}) Integer adminId
     );
 
@@ -350,88 +230,6 @@ public interface AdminShopApi {
     @Operation(summary = "상점 카테고리 삭제")
     @DeleteMapping("/admin/shops/categories/{id}")
     ResponseEntity<Void> deleteShopCategory(
-        @Parameter(in = PATH) @PathVariable Integer id,
-        @Auth(permit = {ADMIN}) Integer adminId
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
-        }
-    )
-    @Operation(summary = "특정 상점의 메뉴카테고리 삭제")
-    @DeleteMapping("/admin/shops/{shopId}/menus/categories/{categoryId}")
-    ResponseEntity<Void> deleteMenuCategory(
-        @Parameter(in = PATH) @PathVariable("shopId") Integer shopId,
-        @Parameter(in = PATH) @PathVariable("categoryId") Integer categoryId,
-        @Auth(permit = {ADMIN}) Integer adminId
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
-        }
-    )
-    @Operation(summary = "특정 상점의 메뉴 삭제")
-    @DeleteMapping("/admin/shops/{shopId}/menus/{menuId}")
-    ResponseEntity<Void> deleteMenu(
-        @Parameter(in = PATH) @PathVariable("shopId") Integer shopId,
-        @Parameter(in = PATH) @PathVariable("menuId") Integer menuId,
-        @Auth(permit = {ADMIN}) Integer adminId
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
-        }
-    )
-    @Operation(summary = "리뷰 조회")
-    @GetMapping("/admin/shops/reviews")
-    ResponseEntity<AdminShopsReviewsResponse> getReviews(
-        @RequestParam(name = "page", defaultValue = "1") Integer page,
-        @RequestParam(name = "limit", defaultValue = "10", required = false) Integer limit,
-        @RequestParam(name = "is_reported", required = false) Boolean isReported,
-        @RequestParam(name = "is_have_unhandled_report", required = false) Boolean isHaveUnhandledReport,
-        @RequestParam(name = "shop_id", required = false) Integer shopId,
-        @Auth(permit = {ADMIN}) Integer adminId
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
-        }
-    )
-    @Operation(summary = "리뷰 신고 상태 변경")
-    @PutMapping("/admin/shops/reviews/{id}")
-    ResponseEntity<Void> modifyReviewReportStatus(
-        @Parameter(in = PATH) @PathVariable Integer id,
-        @RequestBody @Valid AdminModifyShopReviewReportStatusRequest adminModifyShopReviewReportStatusRequest,
-        @Auth(permit = {ADMIN}) Integer adminId
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
-        }
-    )
-    @Operation(summary = "리뷰 삭제")
-    @DeleteMapping("/admin/shops/reviews/{id}")
-    ResponseEntity<Void> deleteReview(
         @Parameter(in = PATH) @PathVariable Integer id,
         @Auth(permit = {ADMIN}) Integer adminId
     );
