+++ src/main/java/in/koreatech/koin/admin/shop/controller/AdminShopController.java
@@ -16,27 +16,17 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import in.koreatech.koin.admin.shop.dto.AdminCreateMenuCategoryRequest;
-import in.koreatech.koin.admin.shop.dto.AdminCreateMenuRequest;
-import in.koreatech.koin.admin.shop.dto.AdminCreateShopCategoryRequest;
-import in.koreatech.koin.admin.shop.dto.AdminCreateShopRequest;
-import in.koreatech.koin.admin.shop.dto.AdminMenuCategoriesResponse;
-import in.koreatech.koin.admin.shop.dto.AdminMenuDetailResponse;
-import in.koreatech.koin.admin.shop.dto.AdminModifyMenuCategoryRequest;
-import in.koreatech.koin.admin.shop.dto.AdminModifyMenuRequest;
-import in.koreatech.koin.admin.shop.dto.AdminModifyShopCategoriesOrderRequest;
-import in.koreatech.koin.admin.shop.dto.AdminModifyShopCategoryRequest;
-import in.koreatech.koin.admin.shop.dto.AdminModifyShopRequest;
-import in.koreatech.koin.admin.shop.dto.AdminModifyShopReviewReportStatusRequest;
-import in.koreatech.koin.admin.shop.dto.AdminShopCategoryResponse;
-import in.koreatech.koin.admin.shop.dto.AdminShopMenuResponse;
-import in.koreatech.koin.admin.shop.dto.AdminShopParentCategoryResponse;
-import in.koreatech.koin.admin.shop.dto.AdminShopResponse;
-import in.koreatech.koin.admin.shop.dto.AdminShopsResponse;
-import in.koreatech.koin.admin.shop.dto.AdminShopsReviewsResponse;
+import in.koreatech.koin.admin.shop.dto.shop.AdminCreateShopCategoryRequest;
+import in.koreatech.koin.admin.shop.dto.shop.AdminCreateShopRequest;
+import in.koreatech.koin.admin.shop.dto.shop.AdminModifyShopCategoriesOrderRequest;
+import in.koreatech.koin.admin.shop.dto.shop.AdminModifyShopCategoryRequest;
+import in.koreatech.koin.admin.shop.dto.shop.AdminModifyShopRequest;
+import in.koreatech.koin.admin.shop.dto.shop.AdminShopCategoryResponse;
+import in.koreatech.koin.admin.shop.dto.shop.AdminShopParentCategoryResponse;
+import in.koreatech.koin.admin.shop.dto.shop.AdminShopResponse;
+import in.koreatech.koin.admin.shop.dto.shop.AdminShopsResponse;
 import in.koreatech.koin.admin.shop.service.AdminShopService;
 import in.koreatech.koin.global.auth.Auth;
-import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -88,124 +78,55 @@ public class AdminShopController implements AdminShopApi {
     public ResponseEntity<List<AdminShopParentCategoryResponse>> getShopParentCategories(
         @Auth(permit = {ADMIN}) Integer adminId
     ) {
-        List<AdminShopParentCategoryResponse> responses = adminShopService.getShopParentCategories();
-        return ResponseEntity.ok(responses);
-    }
-
-    @GetMapping("/admin/shops/{id}/menus")
-    public ResponseEntity<AdminShopMenuResponse> getAllMenus(
-        @Parameter(in = PATH) @PathVariable("id") Integer shopId,
-        @Auth(permit = {ADMIN}) Integer adminId
-    ) {
-        AdminShopMenuResponse adminShopMenuResponse = adminShopService.getAllMenus(shopId);
-        return ResponseEntity.ok(adminShopMenuResponse);
-    }
-
-    @GetMapping("/admin/shops/{id}/menus/categories")
-    public ResponseEntity<AdminMenuCategoriesResponse> getAllMenuCategories(
-        @Parameter(in = PATH) @PathVariable("id") Integer shopId,
-        @Auth(permit = {ADMIN}) Integer adminId
-    ) {
-        AdminMenuCategoriesResponse adminMenuCategoriesResponse = adminShopService.getAllMenuCategories(shopId);
-        return ResponseEntity.ok(adminMenuCategoriesResponse);
-    }
-
-    @GetMapping("/admin/shops/{shopId}/menus/{menuId}")
-    public ResponseEntity<AdminMenuDetailResponse> getMenu(
-        @Parameter(in = PATH) @PathVariable("shopId") Integer shopId,
-        @Parameter(in = PATH) @PathVariable("menuId") Integer menuId,
-        @Auth(permit = {ADMIN}) Integer adminId
-    ) {
-        AdminMenuDetailResponse adminMenuDetailResponse = adminShopService.getMenu(shopId, menuId);
-        return ResponseEntity.ok(adminMenuDetailResponse);
+        List<AdminShopParentCategoryResponse> response = adminShopService.getShopParentCategories();
+        return ResponseEntity.ok(response);
     }
 
     @PostMapping("/admin/shops")
     public ResponseEntity<Void> createShop(
-        @RequestBody @Valid AdminCreateShopRequest adminCreateShopRequest,
+        @RequestBody @Valid AdminCreateShopRequest request,
         @Auth(permit = {ADMIN}) Integer adminId
     ) {
-        adminShopService.createShop(adminCreateShopRequest);
+        adminShopService.createShop(request);
         return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 
     @PostMapping("/admin/shops/categories")
     public ResponseEntity<Void> createShopCategory(
-        @RequestBody @Valid AdminCreateShopCategoryRequest adminCreateShopCategoryRequest,
+        @RequestBody @Valid AdminCreateShopCategoryRequest request,
         @Auth(permit = {ADMIN}) Integer adminId
     ) {
-        adminShopService.createShopCategory(adminCreateShopCategoryRequest);
+        adminShopService.createShopCategory(request);
         return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 
     @PutMapping("/admin/shops/{id}")
     public ResponseEntity<Void> modifyShop(
         @Parameter(in = PATH) @PathVariable Integer id,
-        @RequestBody @Valid AdminModifyShopRequest adminModifyShopRequest,
+        @RequestBody @Valid AdminModifyShopRequest request,
         @Auth(permit = {ADMIN}) Integer adminId
     ) {
-        adminShopService.modifyShop(id, adminModifyShopRequest);
+        adminShopService.modifyShop(id, request);
         return ResponseEntity.ok().build();
     }
 
     @PutMapping("/admin/shops/categories/{id}")
     public ResponseEntity<Void> modifyShopCategory(
         @Parameter(in = PATH) @PathVariable Integer id,
-        @RequestBody @Valid AdminModifyShopCategoryRequest adminModifyShopCategoryRequest,
+        @RequestBody @Valid AdminModifyShopCategoryRequest request,
         @Auth(permit = {ADMIN}) Integer adminId
     ) {
-        adminShopService.modifyShopCategory(id, adminModifyShopCategoryRequest);
+        adminShopService.modifyShopCategory(id, request);
         return ResponseEntity.ok().build();
     }
 
     @PutMapping("/admin/shops/categories/order")
     public ResponseEntity<Void> modifyShopCategoriesOrder(
-        @RequestBody @Valid AdminModifyShopCategoriesOrderRequest adminModifyShopCategoriesOrderRequest,
+        @RequestBody @Valid AdminModifyShopCategoriesOrderRequest request,
         @Auth(permit = {ADMIN}) Integer adminId
     ) {
-        adminShopService.modifyShopCategoriesOrder(adminModifyShopCategoriesOrderRequest);
+        adminShopService.modifyShopCategoriesOrder(request);
         return ResponseEntity.noContent().build();
-    }
-
-    @PutMapping("/admin/shops/{shopId}/menus/categories")
-    public ResponseEntity<Void> modifyMenuCategory(
-        @Parameter(in = PATH) @PathVariable("shopId") Integer shopId,
-        @RequestBody @Valid AdminModifyMenuCategoryRequest adminModifyMenuCategoryRequest,
-        @Auth(permit = {ADMIN}) Integer adminId
-    ) {
-        adminShopService.modifyMenuCategory(shopId, adminModifyMenuCategoryRequest);
-        return ResponseEntity.status(HttpStatus.CREATED).build();
-    }
-
-    @PutMapping("/admin/shops/{shopId}/menus/{menuId}")
-    public ResponseEntity<Void> modifyMenu(
-        @Parameter(in = PATH) @PathVariable("shopId") Integer shopId,
-        @Parameter(in = PATH) @PathVariable("menuId") Integer menuId,
-        @RequestBody @Valid AdminModifyMenuRequest adminModifyMenuRequest,
-        @Auth(permit = {ADMIN}) Integer adminId
-    ) {
-        adminShopService.modifyMenu(shopId, menuId, adminModifyMenuRequest);
-        return ResponseEntity.status(HttpStatus.CREATED).build();
-    }
-
-    @PostMapping("/admin/shops/{id}/menus")
-    public ResponseEntity<Void> createMenu(
-        @Parameter(in = PATH) @PathVariable("id") Integer shopId,
-        @RequestBody @Valid AdminCreateMenuRequest adminCreateMenuRequest,
-        @Auth(permit = {ADMIN}) Integer adminId
-    ) {
-        adminShopService.createMenu(shopId, adminCreateMenuRequest);
-        return ResponseEntity.status(HttpStatus.CREATED).build();
-    }
-
-    @PostMapping("/admin/shops/{id}/menus/categories")
-    public ResponseEntity<Void> createMenuCategory(
-        @Parameter(in = PATH) @PathVariable("id") Integer shopId,
-        @RequestBody @Valid AdminCreateMenuCategoryRequest adminCreateMenuCategoryRequest,
-        @Auth(permit = {ADMIN}) Integer adminId
-    ) {
-        adminShopService.createMenuCategory(shopId, adminCreateMenuCategoryRequest);
-        return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 
     @PostMapping("/admin/shops/{id}/undelete")
@@ -233,62 +154,5 @@ public class AdminShopController implements AdminShopApi {
     ) {
         adminShopService.deleteShopCategory(id);
         return ResponseEntity.noContent().build();
-    }
-
-    @DeleteMapping("/admin/shops/{shopId}/menus/categories/{categoryId}")
-    public ResponseEntity<Void> deleteMenuCategory(
-        @Parameter(in = PATH) @PathVariable("shopId") Integer shopId,
-        @Parameter(in = PATH) @PathVariable("categoryId") Integer categoryId,
-        @Auth(permit = {ADMIN}) Integer adminId
-    ) {
-        adminShopService.deleteMenuCategory(shopId, categoryId);
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-    }
-
-    @DeleteMapping("/admin/shops/{shopId}/menus/{menuId}")
-    public ResponseEntity<Void> deleteMenu(
-        @Parameter(in = PATH) @PathVariable("shopId") Integer shopId,
-        @Parameter(in = PATH) @PathVariable("menuId") Integer menuId,
-        @Auth(permit = {ADMIN}) Integer adminId
-    ) {
-        adminShopService.deleteMenu(shopId, menuId);
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-    }
-
-    @GetMapping("/admin/shops/reviews")
-    public ResponseEntity<AdminShopsReviewsResponse> getReviews(
-        @RequestParam(name = "page", defaultValue = "1") Integer page,
-        @RequestParam(name = "limit", defaultValue = "10", required = false) Integer limit,
-        @RequestParam(name = "is_reported", required = false) Boolean isReported,
-        @RequestParam(name = "has_unhandled_report", required = false) Boolean hasUnhandledReport,
-        @RequestParam(name = "shop_id", required = false) Integer shopId,
-        @Auth(permit = {ADMIN}) Integer adminId
-    ) {
-        return ResponseEntity
-            .status(HttpStatus.OK)
-            .body(adminShopService.getReviews(
-                page, limit, isReported, hasUnhandledReport, shopId
-            ));
-    }
-
-    @Operation(summary = "리뷰 신고 상태 변경")
-    @PutMapping("/admin/shops/reviews/{id}")
-    public ResponseEntity<Void> modifyReviewReportStatus(
-        @Parameter(in = PATH) @PathVariable Integer id,
-        @RequestBody @Valid AdminModifyShopReviewReportStatusRequest adminModifyShopReviewReportStatusRequest,
-        @Auth(permit = {ADMIN}) Integer adminId
-    ) {
-        adminShopService.modifyShopReviewReportStatus(id, adminModifyShopReviewReportStatusRequest);
-        return ResponseEntity.status(HttpStatus.OK).build();
-    }
-
-    @Operation(summary = "리뷰 삭제")
-    @DeleteMapping("/admin/shops/reviews/{id}")
-    public ResponseEntity<Void> deleteReview(
-        @Parameter(in = PATH) @PathVariable Integer id,
-        @Auth(permit = {ADMIN}) Integer adminId
-    ) {
-        adminShopService.deleteShopReview(id);
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
 }
