+++ src/main/java/in/koreatech/koin/admin/shop/dto/menu/AdminModifyMenuRequest.java
@@ -1,26 +1,25 @@
-package in.koreatech.koin.admin.shop.dto;
+package in.koreatech.koin.admin.shop.dto.menu;
 
 import static com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.NOT_REQUIRED;
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
 
+import java.util.List;
+
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import in.koreatech.koin.domain.shop.model.menu.Menu;
-import in.koreatech.koin.domain.shop.model.shop.Shop;
+
 import in.koreatech.koin.global.validation.NotBlankElement;
 import in.koreatech.koin.global.validation.SingleMenuPrice;
 import in.koreatech.koin.global.validation.UniqueId;
 import in.koreatech.koin.global.validation.UniqueUrl;
 import io.swagger.v3.oas.annotations.media.Schema;
-import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PositiveOrZero;
 import jakarta.validation.constraints.Size;
-import java.util.List;
 
 @JsonNaming(value = SnakeCaseStrategy.class)
 @SingleMenuPrice
-public record AdminCreateMenuRequest(
+public record AdminModifyMenuRequest(
     @Schema(example = "[1, 2, 3]", description = "선택된 카테고리 고유 id 리스트", requiredMode = REQUIRED)
     @NotNull(message = "카테고리는 필수입니다.")
     @Size(min = 1, message = "최소 한 개의 카테고리가 필요합니다.")
@@ -33,7 +32,7 @@ public record AdminCreateMenuRequest(
 
     @Schema(example = """
         [ "https://static.koreatech.in/example.png" ]
-        """, description = "이미지 URL 리스트", requiredMode = REQUIRED)
+        """, description = "이미지 URL 리스트", requiredMode = NOT_REQUIRED)
     @Size(max = 3, message = "이미지는 최대 3개까지 입력 가능합니다.")
     @UniqueUrl(message = "이미지 URL은 중복될 수 없습니다.")
     @NotNull(message = "이미지 URL은 null일 수 없습니다.")
@@ -44,7 +43,7 @@ public record AdminCreateMenuRequest(
     @NotNull(message = "단일 메뉴 여부는 필수입니다.")
     boolean isSingle,
 
-    @Schema(example = "짜장면", description = "메뉴명")
+    @Schema(example = "짜장면", description = "메뉴명", requiredMode = REQUIRED)
     @NotNull(message = "메뉴명은 필수입니다.")
     @Size(min = 1, max = 25, message = "메뉴명은 1자 이상 25자 이하로 입력해주세요.")
     String name,
@@ -57,23 +56,17 @@ public record AdminCreateMenuRequest(
     Integer singlePrice
 ) {
 
-    public Menu toEntity(Shop shop) {
-        return Menu.builder()
-            .name(name)
-            .shop(shop)
-            .description(description)
-            .build();
-    }
-
     @JsonNaming(value = SnakeCaseStrategy.class)
-    @Valid
     public record InnerOptionPrice(
         @Schema(example = "대", description = "옵션명", requiredMode = REQUIRED)
-        @NotNull @Size(min = 1, max = 50) String option,
+        @NotNull(message = "옵션명은 필수입니다.")
+        @Size(min = 1, max = 50, message = "옵션명은 1자 이상 50자 이하로 입력해주세요.")
+        String option,
 
         @Schema(example = "26000", description = "가격", requiredMode = REQUIRED)
+        @NotNull(message = "가격은 필수입니다.")
         @PositiveOrZero(message = "가격은 0원 이상이어야 합니다.")
-        @NotNull Integer price
+        Integer price
     ) {
 
     }
