+++ src/main/java/in/koreatech/koin/admin/shop/dto/shop/AdminCreateShopCategoryRequest.java
@@ -1,10 +1,12 @@
-package in.koreatech.koin.admin.shop.dto;
+package in.koreatech.koin.admin.shop.dto.shop;
 
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
+import in.koreatech.koin.domain.shop.model.shop.ShopCategory;
+import in.koreatech.koin.domain.shop.model.shop.ShopParentCategory;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.media.Schema.RequiredMode;
 import jakarta.validation.constraints.NotBlank;
@@ -12,10 +14,10 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 
 @JsonNaming(SnakeCaseStrategy.class)
-public record AdminModifyShopCategoryRequest(
+public record AdminCreateShopCategoryRequest(
     @Schema(description = "이미지 URL", example = "https://static.koreatech.in/test.png", requiredMode = RequiredMode.REQUIRED)
     @NotBlank(message = "이미지 URL은 필수입니다.")
-    @Size(max = 100, message = "이미지 URL은 255자 이하로 입력해주세요.")
+    @Size(max = 255, message = "이미지 URL은 255자 이하로 입력해주세요.")
     String imageUrl,
 
     @Schema(description = "이름", example = "햄버거", requiredMode = RequiredMode.REQUIRED)
@@ -32,4 +34,14 @@ public record AdminModifyShopCategoryRequest(
     String eventBannerImageUrl
 ) {
 
+    public ShopCategory toShopCategory(Integer orderIndex, ShopParentCategory shopParentCategory) {
+        return ShopCategory.builder()
+            .imageUrl(imageUrl)
+            .name(name)
+            .orderIndex(orderIndex)
+            .parentCategory(shopParentCategory)
+            .eventBannerImageUrl(eventBannerImageUrl)
+            .build();
+    }
 }
+
