+++ src/main/java/in/koreatech/koin/admin/shop/dto/shop/AdminModifyShopCategoryRequest.java
@@ -1,12 +1,10 @@
-package in.koreatech.koin.admin.shop.dto;
+package in.koreatech.koin.admin.shop.dto.shop;
 
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
-import in.koreatech.koin.domain.shop.model.shop.ShopCategory;
-import in.koreatech.koin.domain.shop.model.shop.ShopParentCategory;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.media.Schema.RequiredMode;
 import jakarta.validation.constraints.NotBlank;
@@ -14,10 +12,10 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 
 @JsonNaming(SnakeCaseStrategy.class)
-public record AdminCreateShopCategoryRequest(
+public record AdminModifyShopCategoryRequest(
     @Schema(description = "이미지 URL", example = "https://static.koreatech.in/test.png", requiredMode = RequiredMode.REQUIRED)
     @NotBlank(message = "이미지 URL은 필수입니다.")
-    @Size(max = 255, message = "이미지 URL은 255자 이하로 입력해주세요.")
+    @Size(max = 100, message = "이미지 URL은 255자 이하로 입력해주세요.")
     String imageUrl,
 
     @Schema(description = "이름", example = "햄버거", requiredMode = RequiredMode.REQUIRED)
@@ -34,14 +32,4 @@ public record AdminCreateShopCategoryRequest(
     String eventBannerImageUrl
 ) {
 
-    public ShopCategory toShopCategory(Integer orderIndex, ShopParentCategory shopParentCategory) {
-        return ShopCategory.builder()
-            .imageUrl(imageUrl)
-            .name(name)
-            .orderIndex(orderIndex)
-            .parentCategory(shopParentCategory)
-            .eventBannerImageUrl(eventBannerImageUrl)
-            .build();
-    }
 }
-
