+++ src/main/java/in/koreatech/koin/admin/shop/service/AdminShopService.java
@@ -12,7 +12,24 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-import in.koreatech.koin.admin.shop.dto.*;
+import in.koreatech.koin.admin.shop.dto.menu.AdminCreateMenuCategoryRequest;
+import in.koreatech.koin.admin.shop.dto.menu.AdminCreateMenuRequest;
+import in.koreatech.koin.admin.shop.dto.menu.AdminMenuCategoriesResponse;
+import in.koreatech.koin.admin.shop.dto.menu.AdminMenuDetailResponse;
+import in.koreatech.koin.admin.shop.dto.menu.AdminModifyMenuCategoryRequest;
+import in.koreatech.koin.admin.shop.dto.menu.AdminModifyMenuRequest;
+import in.koreatech.koin.admin.shop.dto.menu.AdminShopMenuResponse;
+import in.koreatech.koin.admin.shop.dto.review.AdminModifyShopReviewReportStatusRequest;
+import in.koreatech.koin.admin.shop.dto.review.AdminShopsReviewsResponse;
+import in.koreatech.koin.admin.shop.dto.shop.AdminCreateShopCategoryRequest;
+import in.koreatech.koin.admin.shop.dto.shop.AdminCreateShopRequest;
+import in.koreatech.koin.admin.shop.dto.shop.AdminModifyShopCategoriesOrderRequest;
+import in.koreatech.koin.admin.shop.dto.shop.AdminModifyShopCategoryRequest;
+import in.koreatech.koin.admin.shop.dto.shop.AdminModifyShopRequest;
+import in.koreatech.koin.admin.shop.dto.shop.AdminShopCategoryResponse;
+import in.koreatech.koin.admin.shop.dto.shop.AdminShopParentCategoryResponse;
+import in.koreatech.koin.admin.shop.dto.shop.AdminShopResponse;
+import in.koreatech.koin.admin.shop.dto.shop.AdminShopsResponse;
 import in.koreatech.koin.admin.shop.exception.ShopCategoryNotEmptyException;
 import in.koreatech.koin.admin.shop.exception.ShopCategoryIllegalArgumentException;
 import in.koreatech.koin.admin.shop.repository.*;
@@ -24,8 +41,8 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import in.koreatech.koin.admin.shop.dto.AdminCreateShopRequest.InnerShopOpen;
-import in.koreatech.koin.admin.shop.dto.AdminModifyMenuRequest.InnerOptionPrice;
+import in.koreatech.koin.admin.shop.dto.shop.AdminCreateShopRequest.InnerShopOpen;
+import in.koreatech.koin.admin.shop.dto.menu.AdminModifyMenuRequest.InnerOptionPrice;
 import in.koreatech.koin.admin.shop.exception.ShopCategoryDuplicationException;
 import in.koreatech.koin.domain.shop.model.menu.Menu;
 import in.koreatech.koin.domain.shop.model.menu.MenuCategory;
