+++ src/main/java/in/koreatech/koin/domain/shop/model/shop/Shop.java
@@ -19,7 +19,7 @@ import java.util.Set;
 
 import org.hibernate.annotations.Where;
 
-import in.koreatech.koin.admin.shop.dto.AdminModifyShopRequest;
+import in.koreatech.koin.admin.shop.dto.shop.AdminModifyShopRequest;
 import in.koreatech.koin.domain.owner.model.Owner;
 import in.koreatech.koin.domain.shop.dto.shop.ModifyShopRequest.InnerShopOpen;
 import in.koreatech.koin.domain.shop.model.article.EventArticle;
